pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vstock-data"
version = "1.0.0"
authors = [
  { name="Hunter Do", email="phongvu2010@gmail.com" }
]
description = "Thư viện Python tải dữ liệu chứng khoán Việt Nam từ nhiều nguồn khác nhau."
readme = "README.md"
requires-python = ">=3.10"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial :: Investment"
]
# Các thư viện cốt lõi, luôn được cài đặt
dependencies = ["pandas", "requests"]

# Các nhóm thư viện tùy chọn
[project.optional-dependencies]
yfinance = ["yfinance"]
bigquery = ["google-cloud-bigquery"]
all = ["google-cloud-bigquery", "yfinance"]
dev = ["pytest", "pytest-mock"]

[project.urls]
"Homepage" = "https://github.com/phongvu2010/vStock-Data"
"Bug Tracker" = "https://github.com/phongvu2010/vStock-Data/issues"

[tool.setuptools.packages.find]
where = ["src"]  # Tìm các package trong thư mục "src"

[tool.pytest.ini_options]
pythonpath = ["src"]
